dev: almost implemented milestone 0.1.0.
issue #1 and #3: we have a Schema class with an add_column method that is used
to describe the columns of the CSV with their name and type. The names
make up the header.

issue #9: the main script has mutually exclusive options for number of
rows and number of bytes. It can also be operated in a stream mode which
simply produces data ad infinitum... beware!