ref: renamed add_column to define_column.
The current add_column is a high-level method that might or might not
create an arbitrary or a transformer, does various validation etc.

This is different from the other add_* methods that only do simple
validation and just add the specified item.

Hence define_column will be a "facade" method that does all this stuff,
but I will add a new add_column that simply creates a column with the
given name and does minimal validation.