dev: implemented first proper version of extensible library support.
Addresses #24. It is now possible to define a `Library` object which can
provide arbitrary types, configuration and the compatibility.
Currently it is not possible to actually load a custom library but
support for that will be easy.

Closes #20. It was already completed but we also extended the notion of
compatibility.

Closes #21. It is now possible to specify a global configuration for the
given expression using a `--global-configuration` option.

Closes #13. It is now possible to specify the module to import instead
of `random` by using the `--random-module` option.